+++ src/JumpMenu/Menus/Preview.tsx
@@ -1,9 +1,10 @@
 import React, { useEffect } from 'react'
+import { Heading } from 'rebass'
 
 import PlayerPrimitive from 'Player/PlayerPrimitive'
 import { useVolumeContext, PLAYERS } from 'Player/VolumeContextProvider'
 
-import { useInputContext } from './InputContextProvider'
+import { useInputContext } from '../InputContextProvider'
 
 const Preview: React.FC = () => {
   const { youtubePreviewId } = useInputContext()
@@ -15,13 +16,16 @@ const Preview: React.FC = () => {
   })
 
   return (
-    <PlayerPrimitive
-      playedAt=""
-      youtubeId={youtubePreviewId}
-      controls={true}
-      playerIdentifier={PLAYERS.preview}
-      inline={true}
-    />
+    <>
+      <Heading>Preview</Heading>
+      <PlayerPrimitive
+        playedAt=""
+        youtubeId={youtubePreviewId}
+        controls={true}
+        playerIdentifier={PLAYERS.preview}
+        inline={true}
+      />
+    </>
   )
 }
