+++ src/JumpMenu/Menus/RecordsSearch.tsx
@@ -1,11 +1,12 @@
 import React, { useState } from 'react'
 import { useQuery } from '@apollo/react-hooks'
+import { Heading } from 'rebass'
 
-import { useInputContext } from './InputContextProvider'
-import { SearchQuery, SEARCH_QUERY } from './graphql'
-import Results, { deserialize, Result as ResultType } from './Results'
+import { useInputContext } from '../InputContextProvider'
+import { SearchQuery, SEARCH_QUERY } from '../graphql'
+import Results, { deserialize, Result as ResultType } from '../Results'
 
-const MusicboxSearch: React.FC = () => {
+const RecordsSearch: React.FC = () => {
   const [results, setResults] = useState<ResultType[]>([])
   const { input } = useInputContext()
   const { loading } = useQuery<SearchQuery['data'], SearchQuery['vars']>(SEARCH_QUERY, {
@@ -18,7 +19,12 @@ const MusicboxSearch: React.FC = () => {
     return <p>Loading</p>
   }
 
-  return <Results results={results} />
+  return (
+    <>
+      <Heading>{!!input ? `Searching for ${input}` : 'Showing alll songs'}</Heading>
+      <Results results={results} />
+    </>
+  )
 }
 
-export default MusicboxSearch
+export default RecordsSearch
