+++ src/Room/JumpMenuEntry.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Box, Button } from 'rebass'
 
-import { useJumpMenuContext } from 'JumpMenu'
+import { useJumpNavigationContext } from 'JumpMenu'
 
 const JumpMenuEntry: React.FC = () => {
-  const { show } = useJumpMenuContext()
+  const { show } = useJumpNavigationContext()
   return (
     <Box>
       <Button onClick={show}>Open Jump Menu (j)</Button>
