+++ src/Room/Keyboard/Keyboard.tsx
@@ -1,14 +1,12 @@
-import React, { useCallback, useEffect, useState } from 'react'
+import React, { useCallback, useEffect } from 'react'
 import { useMutation } from '@apollo/react-hooks'
-import { Box, Flex } from 'rebass'
 import { useToasts } from 'react-toast-notifications'
-import { Command } from 'react-feather'
 
 import { useCurrentRecordContext } from 'Context'
 
 import { useApprovalContext } from 'Approval'
 import { SongCreateMutation, SONG_CREATE } from './graphql'
-import { useJumpMenuContext } from 'JumpMenu'
+import { useJumpNavigationContext } from 'JumpMenu'
 
 const Keyboard: React.FC = () => {
   const { addToast } = useToasts()
@@ -20,7 +18,7 @@ const Keyboard: React.FC = () => {
 
   const { currentRecord } = useCurrentRecordContext()
   const { incrementApproval } = useApprovalContext()
-  const { show } = useJumpMenuContext()
+  const { show } = useJumpNavigationContext()
 
   const shortcutHandler = useCallback(
     (ev: KeyboardEvent): void => {
@@ -60,9 +58,7 @@ const Keyboard: React.FC = () => {
     return () => window.removeEventListener('keydown', shortcutHandler)
   }, [shortcutHandler])
 
-  return (
-    <></>
-  )
+  return <></>
 }
 
 export default Keyboard
